feat: ValidateTime, findByName Class
feat: ValidateTime, findByName Class

By testcase, findByName is not working so need update.

Reference by yonghwan
Date : 21.07.29

+++ project.b/guest.book/src/main/java/project/b/guest/book/article/Article.java
@@ -26,6 +26,6 @@ public class Article {
     private String name;
     private String content;
     private long time;
-    private Long likes;
-    private Long dislikes;
+    private int likes;
+    private int dislikes;
 }
